ui: improve scan status handling
